requirements.txt
appdirs==1.4.4
asgiref==3.7.2
astor==0.8.1
attrs==23.1.0
Automat==22.10.0
blessed==1.20.0
CacheControl==0.13.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
ConfigArgParse==1.7
constantly==23.10.4
contexttimer==0.3.3
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.6
django-debug-toolbar==4.2.0
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
docutils==0.20.1
drf-extra-fields==3.7.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
filelock==3.13.1
filetype==1.2.0
flake8==6.1.0
gitdb==4.0.11
GitPython==3.1.40
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
lunr==0.6.2
mccabe==0.7.0
msgpack==1.0.7
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pluggy==1.3.0
pycodestyle==2.11.1
pycparser==2.21
pydoctor==23.9.1
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
smmap==5.0.1
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
srsly==2.4.8
toml==0.10.2
tomli==2.0.1
Twisted==23.10.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.2.12
zope.interface==6.1
